Registro
registro

+++ invoice-manager-ui/src/app/pages/administracion/user/user.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { UsersData} from '../../../@core/data/users-data';
+import { UsersData } from '../../../@core/data/users-data';
 import { HttpErrorResponse } from '@angular/common/http';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { User } from '../../../models/user';
@@ -22,13 +22,13 @@ export class UserComponent implements OnInit {
   registerForm: FormGroup;
   submitted = false;
   public user: User;
-  public rolesArray: IHash =       { "PROMOTOR": false, "TESORERIA": false, "OPERADOR": false, "CONTABILIDAD": false, "ADMINISTRADOR": false, "BOVEDA": false };
-  public rolesArrayUpdate: IHash = { "PROMOTOR": false, "TESORERIA": false, "OPERADOR": false, "CONTABILIDAD": false, "ADMINISTRADOR": false, "BOVEDA": false };
- 
+
+  public rolesArrayUpdate: IHash = { "PROMOTOR": false, "TESORERIA": false, "OPERADOR": false, "CONTABILIDAD": false, "ADMINISTRADOR": false, "BOVEDA": false, "CONSULTOR": false, "OPERADOR-B": false, "OPERADOR-C": false };
+
   public errorMessages: string[] = [];
-  public buttontext:string = "Registrar";;
+  public buttontext: string = "Registrar";;
   public mode: number;
-  
+
   public correorandom: string;
   public inputabalible: boolean;
   public filterParams: any = { correo: this.correorandom, activo: 'true', success: '', message: '', id: '*' };
@@ -41,177 +41,152 @@ export class UserComponent implements OnInit {
   ) { }
 
   ngOnInit() {
-    this.user =  new User();
-    //Validators.pattern("^[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,4}$"),
-    //^[a-z0-9]+(\.[_a-z0-9]+)*@[a-z0-9-]+(\.[a-z0-9-]+)*(\.[a-z]{2,15})$"
-    this.registerForm = this.formBuilder.group({
-      email: ['', [Validators.required, Validators.email, Validators.pattern("^[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,4}$"),]],
-    });
+    this.user = new User();
+
 
     this.errorMessages = [];
     this.filterParams.correo = "correoprueva" + (Math.floor(Math.random() * 6000) + 1).toString() + "@gmail.com";
 
     this.route.paramMap.subscribe(route => {
       const id = route.get('id');
-      console.log(id);
       if (id === '*') {
-     //   this.inputabalible = false;
+        this.inputabalible = false;
         this.mode = 0;
-//      this.registerForm.controls['email'].enable();
-        console.log('Crea nuevo usuario');
       } else {
-     // this.registerForm.controls['email'].disable();
-     // this.inputabalible = true;
-      
-      this.mode = 1;
-      this.buttontext = "Actualiza";
-       
-      console.log('Prersentar informacion de:', id);
-        
-      this.userService.getOneUser(+id).subscribe(
+        this.inputabalible = true;
+
+        this.mode = 1;
+        this.buttontext = "Actualiza";
+        this.userService.getOneUser(+id).subscribe(
           user => {
-            console.log("POST Request Roles bien", user);
-            this.filterParams.correo = user.email;
             this.filterParams.activo = user.activo;
-            this.filterParams.id = user.id;
             this.user = user;
             for (let i in user.roles) {
-        
+
               this.rolesArrayUpdate[user.roles[i].role] = true;
-            
+
             }
             this.CheckRoles();
-            
+
           },
           (error: HttpErrorResponse) => this.errorMessages.push(error.error.message
             || `${error.statusText} : ${error.message}`));
       }
     });
+
+
+    this.registerForm = this.formBuilder.group({
+      email: [{ value: this.filterParams.correo, disabled: this.inputabalible }, [Validators.required, Validators.email, Validators.pattern('^[a-z0-9._%+-]+@[a-z0-9.-]+\\.[a-z]{2,4}$')]],
+    });
+
+
   }
 
   get f() { return this.registerForm.controls; }
 
 
   public Registry() {
-    console.log("but  ");
     this.submitted = true;
 
     if (this.registerForm.invalid) {
       return;
     }
-
     this.errorMessages = [];
 
     if (this.mode == 0) {
       this.userService.insertNewUser(this.filterParams.correo, this.filterParams.activo).subscribe(
         data => {
           this.filterParams.success = 'El usuario ha sido creado satisfactoriamente.'
-          console.log("POST Request Registro bien ", data);
-          for (let i in this.rolesArray) {
-
-            console.log("siclo  " + this.rolesArray[i]);
-
-            if (this.rolesArray[i] != false) {
-              console.log("novacio  " + this.rolesArray[i]);
-
+          for (let i in this.rolesArrayUpdate) {
+            if (this.rolesArrayUpdate[i] != false) {
               this.userService.insertRoles(i, data.id).subscribe(
                 data => {
                   this.filterParams.success = 'El usuario ha sido creado satisfactoriamente.'
-                  console.log("POST Request Roles bien " + i, data);
-
                 },
                 (error: HttpErrorResponse) => this.errorMessages.push(error.error.message
                   || `${error.statusText} : ${error.message}`));
-
             }
           }
         }, (error: HttpErrorResponse) => this.errorMessages.push(error.error.message
           || `${error.statusText} : ${error.message}`));
-      console.log("entro 0 ");
     }
     if (this.mode == 1) {
 
-      this.userService.UpdateUser(+this.filterParams.id, this.filterParams.activo).subscribe(
+      this.userService.UpdateUser(+this.user.id, this.filterParams.activo).subscribe(
         data => {
           this.filterParams.success = 'El usuario ha sido actualizado satisfactoriamente.'
           for (let i in this.user.roles) {
-            //console.log("borrando el ", this.user.roles[i].role);
-            if (this.rolesArrayUpdate[this.user.roles[i].role] == false)
+
+            if (this.rolesArrayUpdate[this.user.roles[i].role] == false && !this.user.roles.find(x => x.role == i)) {
 
               this.userService.DeleteRoles(this.user.id, this.user.roles[i].id).subscribe(
                 data => {
                   this.filterParams.success = 'El usuario ha sido actualizado satisfactoriamente.'
-                  console.log("Delet Request Roles Borrados bien " + i, data);
                   this.actialuza();
                 }, (error: HttpErrorResponse) => this.errorMessages.push(error.error.message
                   || `${error.statusText} : ${error.message}`));
+            }
           }
-          for (let i in this.rolesArray) {
+          for (let i in this.rolesArrayUpdate) {
 
-            console.log("siclo  " + this.rolesArray[i]);
 
-            if (this.rolesArray[i] == true) {
-              console.log("novacio  " + this.rolesArray[i]);
+
+            if (this.rolesArrayUpdate[i] == true && !this.user.roles.find(x => x.role == i)) {
+
 
               this.userService.insertRoles(i, data.id).subscribe(
                 data => {
                   this.filterParams.success = 'El usuario ha sido actualizado satisfactoriamente.'
-                  console.log("POST Request Roles bien " + i, data);
+
                   this.actialuza();
                 },
                 (error: HttpErrorResponse) => this.errorMessages.push(error.error.message
                   || `${error.statusText} : ${error.message}`));
-            this.rolesArray[i] = false      
+              this.rolesArrayUpdate[i] = false
             }
           }
-
-          // this.userService.DeleteRoles 
-          console.log("POST Request Registro bien ", data);
-
         }, (error: HttpErrorResponse) => this.errorMessages.push(error.error.message
           || `${error.statusText} : ${error.message}`));
-      console.log("entro 1 " + this.filterParams.id);
+
 
     }
     //this.filterParams.correo = "correoprueva" + (Math.floor(Math.random() * 6000000) + 1).toString() + "@gmail.com";
 
   }
 
-  
-
   toggle(checked: boolean, rol: string) {
-    
-    this.rolesArray[rol] = checked;
+
+    this.rolesArrayUpdate[rol] = checked;
 
     this.CheckRoles();
-   
+
   }
 
 
 
   public CheckRoles() {
-    for (let i in this.rolesArray) {
+    for (let i in this.rolesArrayUpdate) {
 
-      console.log(this.rolesArray[i]);
+      console.log(this.rolesArrayUpdate[i]);
 
     }
   }
 
-  public actialuza(){
+  public actialuza() {
     this.userService.getOneUser(this.user.id).subscribe(
       user => {
-       
-        console.log("POST Request Roles bien", user);
         this.filterParams.correo = user.email;
         this.filterParams.activo = user.activo;
-        this.filterParams.id = user.id;
+
         this.user = user;
+
         for (let i in user.roles) {
-    
+
           this.rolesArrayUpdate[user.roles[i].role] = true;
-  
+
         }
         this.CheckRoles();
+
 
       },
       (error: HttpErrorResponse) => this.errorMessages.push(error.error.message
